chore: rename sumneko_lua -> lua_ls
